Fixed warnings Updated refraction color to be pixel based to remove artifacts from linear sampling Added zero settings file

+++ Engine/Code/Engine/Renderer/MeshUtils.cpp
@@ -102,10 +102,10 @@ void MeshGenerateNormals( std::vector<Vertex_PCUTBN>& vertices, std::vector<uint
 //---------------------------------------------------------------------------------------------------------
 void AverageNormals( Vec3& currentNormal, Vec3 const& normalToAdd, uint& numAdded )
 {
-	currentNormal *= numAdded;
+	currentNormal *= static_cast<float>( numAdded );
 	currentNormal += normalToAdd;
 	numAdded++;
-	currentNormal /= numAdded;
+	currentNormal /= static_cast<float>( numAdded );
 }
